MGS-5581 [#imx-1868] 0008-change-zp-dtype-from-uint32-to-int32
change zp dtype from uint32 to int32

to support asymetric_affine-i8

Reviewed-on: http://192.168.32.113/review/5826
Tested-by: <builder@verisilicon.com>

+++ ovxlib/include/utils/vsi_nn_dtype_util_prv.h
@@ -154,7 +154,7 @@ static inline int32_t fp32_to_affine
     (
     const float  in,
     const float  scale,
-    const uint8_t    zero_point,
+    const int32_t    zero_point,
     const vsi_nn_type_e type
     )
 {
@@ -171,7 +171,7 @@ static inline float affine_to_fp32
     (
     const int32_t    val,
     const float  scale,
-    const uint8_t    zero_point,
+    const int32_t    zero_point,
     const vsi_nn_type_e type
     )
 {
